@@ -1,5 +1,5 @@
 /** eslint-plugin-perfectionist */
-module.exports = {
+export default {
   "perfectionist/sort-imports": [
     "error",
     {
